requirements_news.txt
anthropic>=0.40.0
feedparser>=6.0.11
python-dotenv>=1.0.0
requests>=2.31.0
trafilatura>=1.12.0
